aside panel and enable input
