+++ components/MonkeyUI.js
@@ -175,7 +175,7 @@ class MonkeyUI extends Component {
 					{ this.props.userSession
 						? ( <div id='mky-content-app' className=''>
 								{ this.state.showConversations & this.haveConversations
-									? <ContentAside isLoadingConversations={this.props.isLoadingConversations} handleLoadMoreConversations={this.props.onLoadMoreConversations} handleConversationDelete={this.props.onConversationDelete} togglePopup={this.togglePopup} handleConversationExit={this.props.onConversationExit} userSessionLogout={this.props.onUserSessionLogout} conversations={this.state.conversations} handleConversationSelected={this.handleConversationSelected} conversationSelected={this.state.conversation} showBanner={this.state.showBanner} show={this.showListConversation} isMobile={this.state.isMobile} closeSide={this.openSide} conversationsLoading={this.props.conversationsLoading}/>
+									? <ContentAside asidePanelParams={this.props.asidePanelParams} connectionStatus={this.props.connectionStatus} isLoadingConversations={this.props.isLoadingConversations} handleLoadMoreConversations={this.props.onLoadMoreConversations} handleConversationDelete={this.props.onConversationDelete} togglePopup={this.togglePopup} handleConversationExit={this.props.onConversationExit} userSessionLogout={this.props.onUserSessionLogout} conversations={this.state.conversations} handleConversationSelected={this.handleConversationSelected} conversationSelected={this.state.conversation} showBanner={this.state.showBanner} show={this.showListConversation} isMobile={this.state.isMobile} closeSide={this.openSide} conversationsLoading={this.props.conversationsLoading}/>
 									: null
 								}
 								<ContentWindow connectionStatus={this.props.connectionStatus} handleNotifyTyping={this.handleNotifyTyping} panelParams={this.props.panelParams} loadMessages={this.props.onMessagesLoad} conversationSelected={this.state.conversation} conversationClosed={this.props.onConversationClosed} messageCreated={this.handleMessageCreated} expandWindow={this.expandWindow} expandAside={this.handleShowAside} isMobile={this.state.isMobile} isPartialized={this.classContent} showBanner={this.state.showBanner} onClickMessage={this.props.onClickMessage} dataDownloadRequest={this.props.onMessageDownloadData} getUser={this.props.onMessageGetUser} haveConversations={this.haveConversations}/>
