+++ styles/chat.css
@@ -818,6 +818,9 @@ TABLE OF CONTENTS
   margin: 0;
   float: left;
   border-right: 1px solid rgba(219,219,219,0.7);
+
+  display: flex;
+  flex-direction: column;
 /*   min-width: 240px; */
 }
 
@@ -832,7 +835,7 @@ TABLE OF CONTENTS
 }
 
 #mky-session-header{
-
+  display: flex;
 }
 
 #mky-expand-each-screen{
@@ -898,6 +901,9 @@ TABLE OF CONTENTS
 .mky-session-conversations{
 	height: calc(100% - 65px);
   position: relative;
+
+  display: flex;
+  flex-direction: column;
 }
 
 .mky-search-input {
@@ -1034,13 +1040,21 @@ TABLE OF CONTENTS
      ----------------------------- */
 
 #mky-conversation-list {
-    width: 100%;
-    height: calc(100% - 51px);
-    margin: 0;
-    padding: 0;
-    list-style: none;
-    overflow: auto;
-    background: #fff;
+  width: 100%;
+  /* height: 100%; */
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  overflow: auto;
+  background: #fff;
+  position: relative;
+  /* display: table-row; */
+  /* position: relative; */
+  display: flex;
+  /* overflow: scroll; */
+  flex: 1;
+  flex-direction: column;
+  -webkit-box-flex: 1;
 }
 
 #mky-conversation-list li{
@@ -1164,6 +1178,35 @@ TABLE OF CONTENTS
   top: 6px;
   right: 0px;
   position: absolute;
+}
+
+.mky-connection-div{
+  width: 100%;
+  background: red;
+  color: white;
+  height: auto;
+  font-weight: bold;
+  -webkit-transition: max-height 2s ease, background-color 0.5s ease;
+  -moz-transition: max-height 2s ease, background-color 0.5s ease;
+  -o-transition: max-height 2s ease, background-color 0.5s ease;
+  transition: max-height 2s ease, background-color 0.5s ease;
+}
+
+.mky-connection-content{
+  margin: 0 auto;
+  padding: 12px 10px 10px 10px;
+}
+
+.mky-connection-content p{
+  margin: 0.2em 0 0 0;
+}
+
+.mky-connection-height{
+  max-height: 200px;
+}
+
+.mky-connection-hide{
+  max-height: 0;
 }
 
 .mky-conversation-image{
@@ -1368,7 +1411,6 @@ TABLE OF CONTENTS
 }
 
 #mky-conversation-selected-header div{
-    display: inline;
     vertical-align: top;
     font-size: 13px;
 }
